🦄 refactor(studio):
- Refactored LinkLocalDependencyButton
- Created ComponentListCard to display all the components
